Refactor Explore component structure
- Updated import path for Explore component to include subdirectory.
- Removed unused Explore component file to clean up the codebase.
