+++ client/src/components/Explore/index.jsx
@@ -1,0 +1,1 @@
+export { default as Explore } from "./Explore"
